feat: Replace prettier-standard with prettier.
The idea here is to create a greater distinction between _formatting_ and _linting_. Formatting should happen automatically prior to a commit, while linting (errors which usually require manual intervention) can run locally or in CI. I like Standard, but the fact that it tries to handle both roles muddies the waters somewhat. Trying to avoid bikeshedding may be valuable, depending on the team, but sometimes configuration can be valuable.

Side Effects:
- Remove --fix from lint script
- Remove eslint rules that prettier handles from .eslintrc
- Remove format script